refactor: 拆分 stock / etf 视图 + 重组 database 目录
视图层：
- v_{bfq,qfq,hfq}_daily 三个混合视图按 class 拆成 6 个：
  v_stock_{bfq,qfq,hfq} / v_etf_{bfq,qfq,hfq}
- 单视图不再 CASE WHEN sc.class，stock 价 ROUND 2 位、ETF 3 位
- 命名格式 v_<class>_<fq>：class 在前便于 tab-complete 按归属浏览
- 不保留兼容视图，下游引用旧名的 SQL 需要迁移
- Schema 版本暂不动（v4.0），等 v4 发布回撤策略确定再升

代码层：database/{duckdb,clickhouse}/ 7 文件结构替代旧 ddl.go/dml.go：
  driver.go       连接 / 生命周期 / 结构体
  schema.go       纯 schema：建表 + 视图编排 + mapType
  meta.go         _meta 表 I/O：schema 版本 r/w（预留 run 审计日志）
  import.go       ImportCSV / 6 Import* / TruncateTable / RebuildSymbolClass
  query.go        所有 Get* / Query / Count* (纯读)
  views_stock.go  3 个 stock 视图独立 builder + registerStockViews
  views_etf.go    3 个 etf 视图独立 builder + registerETFViews
                  + ETF 规则速查注释（cat=11、精度、basic_daily 字段有效性）

视图 builder 也从"3 个参数化函数"展开成"6 个独立函数"——stock 与 ETF
本质是两个领域，强行 DRY 会阻碍未来分化。
