--- database/clickhouse/ddl.go
@@ -1,195 +1,0 @@
-package clickhouse
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/jing2uo/tdx2db/model"
-)
-
-// mapType 针对 ClickHouse 进行类型优化
-func (d *ClickHouseDriver) mapType(colName string, dt model.DataType) string {
-	isKey := strings.Contains(strings.ToLower(colName), "symbol")
-
-	switch dt {
-	case model.TypeString:
-		if isKey {
-			return "LowCardinality(String)"
-		}
-		return "String"
-	case model.TypeFloat64:
-		return "Float64"
-	case model.TypeInt64:
-		return "Int64"
-	case model.TypeDate:
-		return "Date32"
-	case model.TypeDateTime:
-		return "DateTime64(0, 'Asia/Shanghai')"
-	default:
-		return "String"
-	}
-}
-
-func (d *ClickHouseDriver) createTableInternal(meta *model.TableMeta) error {
-	var colDefs []string
-	var dateCol, keyCol string
-
-	// 1. 构建列定义
-	for _, col := range meta.Columns {
-		sqlType := d.mapType(col.Name, col.Type)
-		colDefs = append(colDefs, fmt.Sprintf("%s %s", col.Name, sqlType))
-
-		// 自动探测用于排序键的列
-		lowerName := strings.ToLower(col.Name)
-		if lowerName == "date" || lowerName == "datetime" {
-			dateCol = col.Name
-		}
-		if lowerName == "symbol" {
-			keyCol = col.Name
-		}
-	}
-
-	// 2. 确定排序键 (MergeTree 必须)
-	orderBy := "tuple()"
-	if keyCol != "" && dateCol != "" {
-		orderBy = fmt.Sprintf("(%s, %s)", keyCol, dateCol)
-	} else if dateCol != "" {
-		orderBy = dateCol
-	} else if keyCol != "" {
-		orderBy = keyCol
-	}
-
-	// 3. 建表语句
-	query := fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			%s
-		) ENGINE = MergeTree()
-		ORDER BY %s
-	`, meta.TableName, strings.Join(colDefs, ", "), orderBy)
-
-	_, err := d.db.Exec(query)
-	return err
-}
-
-func (d *ClickHouseDriver) registerViews() {
-
-	// 不复权视图 (BFQ) — stock + etf
-	d.viewImpls[model.ViewDailyBFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			WITH latest_factors AS (
-				SELECT symbol, argMax(hfq_factor, date) AS latest_hfq
-				FROM %s
-				GROUP BY symbol
-			)
-			SELECT
-				s.symbol     AS symbol,
-				s.date       AS date,
-				s.open       AS open,
-				s.high       AS high,
-				s.low        AS low,
-				s.close      AS close,
-				b.preclose   AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) AS hfq_factor,
-				COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1) AS qfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN latest_factors lf ON s.symbol = lf.symbol
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyBFQ,
-			model.TableAdjustFactor.TableName,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-
-	// 后复权视图 (HFQ) — stock + etf
-	d.viewImpls[model.ViewDailyHFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			SELECT
-				s.symbol   AS symbol,
-				s.date     AS date,
-				ROUND(s.open  * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS open,
-				ROUND(s.high  * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS high,
-				ROUND(s.low   * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS low,
-				ROUND(s.close * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS close,
-				ROUND(b.preclose * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) AS hfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyHFQ,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-
-	// 前复权视图 (QFQ) — stock + etf
-	d.viewImpls[model.ViewDailyQFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			WITH latest_factors AS (
-				SELECT symbol, argMax(hfq_factor, date) AS latest_hfq
-				FROM %s
-				GROUP BY symbol
-			)
-			SELECT
-				s.symbol   AS symbol,
-				s.date     AS date,
-				ROUND(s.open  * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS open,
-				ROUND(s.high  * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS high,
-				ROUND(s.low   * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS low,
-				ROUND(s.close * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS close,
-				ROUND(b.preclose * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1) AS qfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN latest_factors lf ON s.symbol = lf.symbol
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyQFQ,
-			model.TableAdjustFactor.TableName,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-}
