--- database/clickhouse/dml.go
@@ -1,285 +1,0 @@
-package clickhouse
-
-import (
-	"context"
-	"database/sql"
-	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"strings"
-	"time"
-
-	"github.com/jing2uo/tdx2db/model"
-)
-
-func (d *ClickHouseDriver) ImportCSV(meta *model.TableMeta, filePath string) error {
-	file, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	req, err := http.NewRequest("POST", d.httpImportUrl, file)
-	if err != nil {
-		return err
-	}
-
-	req.Header.Set("Content-Type", "text/csv")
-
-	// 设置参数
-	q := req.URL.Query()
-
-	if d.database != "" {
-		q.Set("database", d.database)
-	}
-
-	q.Add("query", fmt.Sprintf("INSERT INTO %s FORMAT CSVWithNames", meta.TableName))
-	q.Add("date_time_input_format", "best_effort")
-	q.Add("session_timezone", "Asia/Shanghai")
-
-	req.URL.RawQuery = q.Encode()
-
-	if d.authUser != "" {
-		req.SetBasicAuth(d.authUser, d.authPass)
-	}
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		bodyBytes, _ := io.ReadAll(resp.Body)
-		errMsg := strings.TrimSpace(string(bodyBytes))
-		return fmt.Errorf("clickhouse insert failed (db: %s, status %d): %s", d.database, resp.StatusCode, errMsg)
-	}
-
-	return nil
-}
-
-func (d *ClickHouseDriver) TruncateTable(meta *model.TableMeta) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	query := fmt.Sprintf("TRUNCATE TABLE IF EXISTS %s", meta.TableName)
-
-	_, err := d.db.ExecContext(ctx, query)
-	if err != nil {
-		return fmt.Errorf("clickhouse truncate via tcp failed: %w", err)
-	}
-
-	return nil
-}
-
-func (d *ClickHouseDriver) ImportKlineDaily(path string) error {
-	return d.ImportCSV(model.TableKlineDaily, path)
-}
-
-func (d *ClickHouseDriver) ImportKline1Min(path string) error {
-	return d.ImportCSV(model.TableKline1Min, path)
-}
-
-func (d *ClickHouseDriver) ImportGBBQ(path string) error {
-	d.TruncateTable(model.TableGbbq)
-	return d.ImportCSV(model.TableGbbq, path)
-}
-
-func (d *ClickHouseDriver) ImportBasic(path string) error {
-	return d.ImportCSV(model.TableBasicDaily, path)
-}
-
-func (d *ClickHouseDriver) ImportAdjustFactors(path string) error {
-	return d.ImportCSV(model.TableAdjustFactor, path)
-}
-
-func (d *ClickHouseDriver) ImportHolidays(path string) error {
-	d.TruncateTable(model.TableHoliday)
-	return d.ImportCSV(model.TableHoliday, path)
-}
-
-func (d *ClickHouseDriver) Query(table string, conditions map[string]interface{}, dest interface{}) error {
-	query := fmt.Sprintf("SELECT * FROM %s", table)
-	args := []interface{}{}
-	if len(conditions) > 0 {
-		whereParts := []string{}
-		for k, v := range conditions {
-			whereParts = append(whereParts, fmt.Sprintf("%s = ?", k))
-			args = append(args, v)
-		}
-		query += " WHERE " + strings.Join(whereParts, " AND ")
-	}
-
-	return d.db.Select(dest, query, args...)
-}
-
-func (d *ClickHouseDriver) GetLatestDate(tableName string, dateCol string) (time.Time, error) {
-	query := fmt.Sprintf("SELECT toDate(maxOrNull(%s)) AS latest FROM %s", dateCol, tableName)
-	var latest sql.NullTime
-	err := d.db.Get(&latest, query)
-	if err != nil {
-		return time.Time{}, err
-	}
-	if !latest.Valid {
-		return time.Time{}, nil
-	}
-	return latest.Time, nil
-}
-
-func (d *ClickHouseDriver) GetSymbolsByClass(classes ...string) ([]string, error) {
-	if len(classes) == 0 {
-		return nil, nil
-	}
-	placeholders := strings.Repeat("?,", len(classes))
-	placeholders = placeholders[:len(placeholders)-1]
-	query := fmt.Sprintf(
-		"SELECT symbol FROM %s WHERE class IN (%s) ORDER BY symbol",
-		model.TableSymbolClass.TableName, placeholders,
-	)
-	args := make([]any, len(classes))
-	for i, c := range classes {
-		args[i] = c
-	}
-
-	var symbols []string
-	if err := d.db.Select(&symbols, query, args...); err != nil {
-		return nil, fmt.Errorf("failed to query symbols by class: %w", err)
-	}
-	return symbols, nil
-}
-
-func (d *ClickHouseDriver) RebuildSymbolClass() error {
-	kline := model.TableKlineDaily.TableName
-	class := model.TableSymbolClass.TableName
-
-	var codes []string
-	if err := d.db.Select(&codes, fmt.Sprintf("SELECT DISTINCT symbol FROM %s", kline)); err != nil {
-		return fmt.Errorf("failed to collect symbols: %w", err)
-	}
-
-	if err := d.TruncateTable(model.TableSymbolClass); err != nil {
-		return err
-	}
-
-	if len(codes) == 0 {
-		return nil
-	}
-
-	var sb strings.Builder
-	sb.WriteString(fmt.Sprintf("INSERT INTO %s (symbol, class) VALUES ", class))
-	for i, c := range codes {
-		if i > 0 {
-			sb.WriteString(", ")
-		}
-		sb.WriteString(fmt.Sprintf("('%s', '%s')", c, model.ClassifyCode(c)))
-	}
-
-	if _, err := d.db.Exec(sb.String()); err != nil {
-		return fmt.Errorf("failed to insert symbol_class: %w", err)
-	}
-	return nil
-}
-
-func (d *ClickHouseDriver) CountKlineDaily() (int64, error) {
-	query := fmt.Sprintf("SELECT COUNT(*) FROM %s", model.TableKlineDaily.TableName)
-
-	var count int64
-	err := d.db.Get(&count, query)
-	if err != nil {
-		return 0, fmt.Errorf("failed to count kline daily: %w", err)
-	}
-
-	return count, nil
-}
-
-func (d *ClickHouseDriver) QueryKlineDaily(symbol string, startDate, endDate *time.Time) ([]model.KlineDay, error) {
-
-	conditions := []string{"symbol = ?"}
-	args := []interface{}{symbol}
-
-	if startDate != nil {
-		conditions = append(conditions, "date >= ?")
-		args = append(args, *startDate)
-	}
-	if endDate != nil {
-		conditions = append(conditions, "date <= ?")
-		args = append(args, *endDate)
-	}
-
-	query := fmt.Sprintf(
-		`SELECT * FROM %s WHERE %s ORDER BY date ASC`,
-		model.TableKlineDaily.TableName,
-		strings.Join(conditions, " AND "),
-	)
-
-	var results []model.KlineDay
-	if err := d.db.Select(&results, query, args...); err != nil {
-		return nil, fmt.Errorf("failed to query kline daily: %w", err)
-	}
-
-	return results, nil
-}
-
-func (d *ClickHouseDriver) GetBasicsBySymbol(symbol string) ([]model.BasicDaily, error) {
-	query := fmt.Sprintf(
-		"SELECT * FROM %s WHERE symbol = ? ORDER BY date",
-		model.TableBasicDaily.TableName,
-	)
-
-	var results []model.BasicDaily
-	if err := d.db.Select(&results, query, symbol); err != nil {
-		return nil, fmt.Errorf("failed to query daily basics by symbol %s: %w", symbol, err)
-	}
-
-	return results, nil
-}
-
-func (d *ClickHouseDriver) GetGbbq() ([]model.GbbqData, error) {
-	table := model.TableGbbq.TableName
-
-	query := fmt.Sprintf(`SELECT * FROM %s ORDER BY symbol, date`, table)
-
-	var results []model.GbbqData
-	if err := d.db.Select(&results, query); err != nil {
-		return nil, fmt.Errorf("failed to query gbbq: %w", err)
-	}
-
-	return results, nil
-}
-
-func (d *ClickHouseDriver) GetHolidays() ([]time.Time, error) {
-	query := fmt.Sprintf("SELECT date FROM %s ORDER BY date", model.TableHoliday.TableName)
-	var dates []time.Time
-	if err := d.db.Select(&dates, query); err != nil {
-		return nil, fmt.Errorf("failed to query holidays: %w", err)
-	}
-	return dates, nil
-}
-
-const chMetaTable = "_meta"
-
-func (d *ClickHouseDriver) ReadSchemaVersion() (string, error) {
-	var value string
-	err := d.db.Get(&value,
-		fmt.Sprintf("SELECT value FROM %s WHERE key = 'schema_version'", chMetaTable))
-	if err == sql.ErrNoRows {
-		return "", nil
-	}
-	if err != nil {
-		return "", fmt.Errorf("failed to read schema version: %w", err)
-	}
-	return value, nil
-}
-
-func (d *ClickHouseDriver) WriteSchemaVersion() error {
-	_, err := d.db.Exec(
-		fmt.Sprintf("INSERT INTO %s (key, value) VALUES ('schema_version', ?)", chMetaTable),
-		fmt.Sprintf("%d.%d", model.SchemaMajor, model.SchemaMinor),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to write schema version: %w", err)
-	}
-	return nil
-}
