+++ database/clickhouse/driver.go
@@ -161,25 +161,3 @@ func (d *ClickHouseDriver) Close() error {
 	}
 	return nil
 }
-
-func (d *ClickHouseDriver) InitSchema() error {
-	tables := model.AllTables()
-
-	for _, t := range tables {
-		if err := d.createTableInternal(t); err != nil {
-			return fmt.Errorf("failed to create table %s: %w", t.TableName, err)
-		}
-	}
-
-	d.registerViews()
-	for _, viewID := range model.AllViews() {
-		implFunc, exists := d.viewImpls[viewID]
-		if !exists {
-			return fmt.Errorf("[ClickHouse] Missing implementation for view: %s", viewID)
-		}
-		if err := implFunc(); err != nil {
-			return fmt.Errorf("failed to create view %s: %w", viewID, err)
-		}
-	}
-	return nil
-}
