--- database/duckdb/ddl.go
@@ -1,167 +1,0 @@
-package duckdb
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/jing2uo/tdx2db/model"
-)
-
-// mapType 将通用 DataType 转换为 DuckDB 的 SQL 类型
-func (d *DuckDBDriver) mapType(dt model.DataType) string {
-	switch dt {
-	case model.TypeString:
-		return "VARCHAR"
-	case model.TypeFloat64:
-		return "DOUBLE"
-	case model.TypeInt64:
-		return "BIGINT"
-	case model.TypeDate:
-		return "DATE"
-	case model.TypeDateTime:
-		return "TIMESTAMPTZ"
-	default:
-		return "VARCHAR"
-	}
-}
-
-func (d *DuckDBDriver) createTableInternal(meta *model.TableMeta) error {
-	var colDefs []string
-	for _, col := range meta.Columns {
-		// 1. 获取 DuckDB 具体的类型
-		sqlType := d.mapType(col.Type)
-		// 2. 拼接字段定义
-		colDefs = append(colDefs, fmt.Sprintf("%s %s", col.Name, sqlType))
-	}
-
-	query := fmt.Sprintf("CREATE TABLE IF NOT EXISTS %s (%s)",
-		meta.TableName, strings.Join(colDefs, ", "))
-
-	_, err := d.db.Exec(query)
-	if err != nil {
-		return fmt.Errorf("failed to create table %s: %w", meta.TableName, err)
-	}
-	return nil
-}
-
-func (d *DuckDBDriver) registerViews() {
-	// 不复权视图 (BFQ) — stock + etf
-	d.viewImpls[model.ViewDailyBFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			WITH latest_factors AS (
-				SELECT symbol, argMax(hfq_factor, date) AS latest_hfq
-				FROM %s
-				GROUP BY symbol
-			)
-			SELECT
-				s.symbol     AS symbol,
-				s.date       AS date,
-				s.open       AS open,
-				s.high       AS high,
-				s.low        AS low,
-				s.close      AS close,
-				b.preclose   AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) AS hfq_factor,
-				COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1) AS qfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN latest_factors lf ON s.symbol = lf.symbol
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyBFQ,
-			model.TableAdjustFactor.TableName,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-
-	// 后复权视图 (HFQ) — stock + etf
-	d.viewImpls[model.ViewDailyHFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			SELECT
-				s.symbol   AS symbol,
-				s.date     AS date,
-				ROUND(s.open  * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS open,
-				ROUND(s.high  * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS high,
-				ROUND(s.low   * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS low,
-				ROUND(s.close * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS close,
-				ROUND(b.preclose * COALESCE(f.hfq_factor, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) AS hfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyHFQ,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-
-	// 前复权视图 (QFQ) — stock + etf
-	d.viewImpls[model.ViewDailyQFQ] = func() error {
-		query := fmt.Sprintf(`
-			CREATE OR REPLACE VIEW %s AS
-			WITH latest_factors AS (
-				SELECT symbol, argMax(hfq_factor, date) AS latest_hfq
-				FROM %s
-				GROUP BY symbol
-			)
-			SELECT
-				s.symbol   AS symbol,
-				s.date     AS date,
-				ROUND(s.open  * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS open,
-				ROUND(s.high  * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS high,
-				ROUND(s.low   * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS low,
-				ROUND(s.close * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS close,
-				ROUND(b.preclose * COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1), CASE WHEN sc.class = 'etf' THEN 3 ELSE 2 END) AS preclose,
-				s.volume     AS volume,
-				s.amount     AS amount,
-				b.turnover   AS turnover,
-				b.floatmv    AS floatmv,
-				b.totalmv    AS totalmv,
-				b.change_pct AS change_pct,
-				b.amplitude  AS amplitude,
-				COALESCE(f.hfq_factor, 1) / COALESCE(lf.latest_hfq, 1) AS qfq_factor
-			FROM %s s
-			INNER JOIN %s sc ON s.symbol = sc.symbol AND sc.class IN ('stock', 'etf')
-			LEFT JOIN %s f ON s.symbol = f.symbol AND s.date = f.date
-			LEFT JOIN latest_factors lf ON s.symbol = lf.symbol
-			LEFT JOIN %s b ON s.symbol = b.symbol AND s.date = b.date
-		`,
-			model.ViewDailyQFQ,
-			model.TableAdjustFactor.TableName,
-			model.TableKlineDaily.TableName,
-			model.TableSymbolClass.TableName,
-			model.TableAdjustFactor.TableName,
-			model.TableBasicDaily.TableName,
-		)
-		_, err := d.db.Exec(query)
-		return err
-	}
-}
