+++ database/duckdb/query.go
@@ -9,67 +9,6 @@ import (
 	"github.com/jing2uo/tdx2db/model"
 )
 
-func (d *DuckDBDriver) ImportCSV(meta *model.TableMeta, csvPath string) error {
-	var colMaps []string
-	for _, col := range meta.Columns {
-		duckType := d.mapType(col.Type)
-		colMaps = append(colMaps, fmt.Sprintf("'%s': '%s'", col.Name, duckType))
-	}
-
-	columnsStr := strings.Join(colMaps, ", ")
-
-	query := fmt.Sprintf(`
-		INSERT INTO %s
-		SELECT * FROM read_csv('%s',
-			header=true,
-			columns={%s},
-			dateformat='%%Y-%%m-%%d',
-			timestampformat='%%Y-%%m-%%d %%H:%%M'
-		)
-	`, meta.TableName, csvPath, columnsStr)
-
-	_, err := d.db.Exec(query)
-	return err
-}
-
-func (d *DuckDBDriver) TruncateTable(meta *model.TableMeta) error {
-
-	query := fmt.Sprintf("DELETE FROM %s", meta.TableName)
-
-	_, err := d.db.Exec(query)
-	if err != nil {
-		return fmt.Errorf("duckdb truncate failed: %w", err)
-	}
-
-	return nil
-}
-
-func (d *DuckDBDriver) ImportKlineDaily(path string) error {
-	return d.ImportCSV(model.TableKlineDaily, path)
-}
-
-func (d *DuckDBDriver) ImportKline1Min(path string) error {
-	return d.ImportCSV(model.TableKline1Min, path)
-}
-
-func (d *DuckDBDriver) ImportGBBQ(path string) error {
-	d.TruncateTable(model.TableGbbq)
-	return d.ImportCSV(model.TableGbbq, path)
-}
-
-func (d *DuckDBDriver) ImportBasic(path string) error {
-	return d.ImportCSV(model.TableBasicDaily, path)
-}
-
-func (d *DuckDBDriver) ImportAdjustFactors(path string) error {
-	return d.ImportCSV(model.TableAdjustFactor, path)
-}
-
-func (d *DuckDBDriver) ImportHolidays(path string) error {
-	d.TruncateTable(model.TableHoliday)
-	return d.ImportCSV(model.TableHoliday, path)
-}
-
 func (d *DuckDBDriver) Query(table string, conditions map[string]interface{}, dest interface{}) error {
 	query := fmt.Sprintf("SELECT * FROM %s", table)
 	args := []interface{}{}
@@ -125,40 +64,6 @@ func (d *DuckDBDriver) GetSymbolsByClass(classes ...string) ([]string, error) {
 	return symbols, nil
 }
 
-func (d *DuckDBDriver) RebuildSymbolClass() error {
-	kline := model.TableKlineDaily.TableName
-	class := model.TableSymbolClass.TableName
-
-	var codes []string
-	if err := d.db.Select(&codes, fmt.Sprintf("SELECT DISTINCT symbol FROM %s", kline)); err != nil {
-		return fmt.Errorf("failed to collect symbols: %w", err)
-	}
-
-	tx, err := d.db.Beginx()
-	if err != nil {
-		return fmt.Errorf("failed to begin tx: %w", err)
-	}
-	defer tx.Rollback()
-
-	if _, err := tx.Exec(fmt.Sprintf("DELETE FROM %s", class)); err != nil {
-		return fmt.Errorf("failed to clear symbol_class: %w", err)
-	}
-
-	stmt, err := tx.Preparex(fmt.Sprintf("INSERT INTO %s (symbol, class) VALUES (?, ?)", class))
-	if err != nil {
-		return fmt.Errorf("failed to prepare insert: %w", err)
-	}
-	defer stmt.Close()
-
-	for _, c := range codes {
-		if _, err := stmt.Exec(c, model.ClassifyCode(c)); err != nil {
-			return fmt.Errorf("failed to insert class for %s: %w", c, err)
-		}
-	}
-
-	return tx.Commit()
-}
-
 func (d *DuckDBDriver) CountKlineDaily() (int64, error) {
 	query := fmt.Sprintf("SELECT COUNT(*) FROM %s", model.TableKlineDaily.TableName)
 
@@ -172,7 +77,6 @@ func (d *DuckDBDriver) CountKlineDaily() (int64, error) {
 }
 
 func (d *DuckDBDriver) QueryKlineDaily(symbol string, startDate, endDate *time.Time) ([]model.KlineDay, error) {
-
 	conditions := []string{"symbol = ?"}
 	args := []interface{}{symbol}
 
@@ -233,30 +137,4 @@ func (d *DuckDBDriver) GetHolidays() ([]time.Time, error) {
 		return nil, fmt.Errorf("failed to query holidays: %w", err)
 	}
 	return dates, nil
-}
-
-const metaTable = "_meta"
-
-func (d *DuckDBDriver) ReadSchemaVersion() (string, error) {
-	var value string
-	err := d.db.Get(&value,
-		fmt.Sprintf("SELECT value FROM %s WHERE key = 'schema_version'", metaTable))
-	if err == sql.ErrNoRows {
-		return "", nil
-	}
-	if err != nil {
-		return "", fmt.Errorf("failed to read schema version: %w", err)
-	}
-	return value, nil
-}
-
-func (d *DuckDBDriver) WriteSchemaVersion() error {
-	_, err := d.db.Exec(
-		fmt.Sprintf("INSERT INTO %s (key, value) VALUES ('schema_version', ?)", metaTable),
-		fmt.Sprintf("%d.%d", model.SchemaMajor, model.SchemaMinor),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to write schema version: %w", err)
-	}
-	return nil
 }
